backend/requirements.txt
# DroneMate Python Dependencies

# Web Framework
fastapi==0.111.0
uvicorn[standard]==0.29.0

# Database
motor==3.4.0
pymongo==4.7.2

# AI & ML
requests==2.31.0
sentence-transformers==2.7.0

# Groq API client (optional - needed only if using Groq backup)
# groq==0.8.0

# Gemini API client (optional - needed only if using Gemini backup)
# google-generativeai==0.5.4

# Utilities
python-dotenv==1.0.1
python-multipart==0.0.9
pydantic==2.7.1

# Text Processing
nltk==3.8.1

# CORS
# (included with fastapi)

# Development
httpx==0.27.0
